Bug 424165 - Drive JSON creation and update tests using OSLC shapes
Use shapes for JSON creation and update tests if no template files are
provided in the test setup properties file. Also refactor the tests to
avoid passing so many parameters. The tests are now defined in
CreationAndUpdateBaseTests.java, and only a few methods need to be
overriden in the sublasses.

Change-Id: I19e3200e33031e3d7759c6882c95054e120de344
Signed-off-by: Samuel Padgett <spadgett@us.ibm.com>